+++ server/src/main/kotlin/com/adobe/testing/s3mock/controller/ControllerConfiguration.kt
@@ -29,9 +29,9 @@ import jakarta.servlet.Filter
 import jakarta.servlet.http.HttpServletRequest
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
+import org.springframework.boot.autoconfigure.condition.ConditionalOnBean
 import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.boot.servlet.filter.OrderedFormContentFilter
-import org.springframework.boot.autoconfigure.condition.ConditionalOnBean
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.context.annotation.Profile